+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 2.1.9-SNAPSHOT / WIP
 ==================
 
+Improvements:
+* [OE-100](https://openlmis.atlassian.net/browse/OE-100): Added quantity-unit-toggle to Stock on Hand page
+
 2.1.8 / 2025-03-31
 ==================
 * [OLMIS-8126](https://openlmis.atlassian.net/browse/OLMIS-8126): Fixed display of expiration date for lots